backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
pytest>=7.4.0
httpx>=0.24.1
alembic>=1.12.0
sentence-transformers>=2.2.2
langchain>=0.0.300
chromadb>=0.4.13
numpy>=1.25.2
pandas>=2.1.0
pydantic[email]>=2.0.0